+++ helpers.js
@@ -126,7 +126,7 @@ export class MEJHelpers {
     static toDefaultCurrency(price) {
         let value = (typeof price == "string" ? MEJHelpers.getPrice(price, "price") : price);
         let currency = MonksEnhancedJournal.currencies.find(c => c.id == value.currency);
-        let result = (currency.convert || 1) * value.value;
+        let result = (currency?.convert || 1) * value.value;
 
         return result;
     }